+++ internal-api/src/main/java/datadog/trace/bootstrap/instrumentation/api/AgentSpan.java
@@ -8,7 +8,7 @@ import datadog.trace.api.interceptor.MutableSpan;
 import datadog.trace.api.sampling.PrioritySampling;
 import java.util.Map;
 
-public interface AgentSpan extends MutableSpan, IGSpanInfo {
+public interface AgentSpan extends MutableSpan, IGSpanInfo, PathwayContextHolder {
 
   DDTraceId getTraceId();
 
@@ -139,8 +139,6 @@ public interface AgentSpan extends MutableSpan, IGSpanInfo {
 
   /** RequestContext for the Instrumentation Gateway */
   RequestContext getRequestContext();
-
-  void mergePathwayContext(PathwayContext pathwayContext);
 
   Integer forceSamplingDecision();
