+++ internal-api/src/main/java/datadog/trace/bootstrap/instrumentation/api/AgentTracer.java
@@ -595,6 +595,11 @@ public class AgentTracer {
     }
 
     @Override
+    public PathwayContext getPathwayContext() {
+      return NoopPathwayContext.INSTANCE;
+    }
+
+    @Override
     public void mergePathwayContext(PathwayContext pathwayContext) {}
 
     @Override
