+++ internal-api/src/main/java/datadog/trace/bootstrap/instrumentation/api/PathwayContextHolder.java
@@ -1,0 +1,7 @@
+package datadog.trace.bootstrap.instrumentation.api;
+
+public interface PathwayContextHolder {
+  PathwayContext getPathwayContext();
+
+  void mergePathwayContext(PathwayContext pathwayContext);
+}
